fix(build): Fix path of EL0 apps linker script
To create the linker scripts of EL0 apps, we copy the common script to
the binary directory then use arm_target_linker_script() which calls
arm_preprocess_source(). At the moment arm_preprocess_source() expects
the input to be in the source directory and doesn't work properly when
it is in the binary directory, because it tries to get a relative path
from CMAKE_CURRENT_SOURCE_DIR to the input, which in this case is in
CMAKE_CURRENT_BINARY_DIR. For example with:

    CMAKE_CURRENT_SOURCE_DIR=/my/source/
    CMAKE_CURRENT_BINARY_DIR=/path/to/build/
    ATTEST_APP_LINKER_SCRIPT=/path/to/build/linker_attest.lds.i

then after:

    file(RELATIVE_PATH preprocessed_source "${CMAKE_CURRENT_SOURCE_DIR}"
                    "${preprocessed_source}")
    string(PREPEND preprocessed_source "${CMAKE_CURRENT_BINARY_DIR}/")

we end up with:

    preprocessed_source=
        /path/to/build/../../path/to/build/linker_attest.lds.i

whose directory doesn't exist, resulting in the error:

    CMake Error at app/attestation/el0_app/CMakeLists.txt:145 (file):
    file failed to create directory:

    /path/to/build/../../path/to/build/linker_attest.lds.i

    because: No such file or directory

Since we preprocess the source script we don't actually need the extra
step of copying the common script, we can just give it as input to the
preprocessor. Pass the common script as argument and rework the output
path calculation.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Change-Id: Iec6ce7f695b1316e4e4efcf5b3b7185b3d5f6aa1