feat(lib): enable access to smc_arch_feature_availability
SMCCC_ARCH_FEATURE_AVAILABILITY is an smc call invoked to query
EL3 about the features supported or enabled in firmware.

Lower exception levels(R-EL2 in this case) assumes that a feature
is available based on hardware capability and may directly read
the corresponding ID registers. However, if a feature is not
enabled or supported in firmware, any access to its associated
registers from a lower EL could lead to traps to EL3 or result
in undefined behavior.

To prevent this, the SMC call provides a mechanism for lower ELs
to help verify feature availability inturn ensuring valid access to
registers for features explicitly supported and enabled by EL3.

For a list of features and their corresponding bitmasks,
please refer to the documentation linked below:

Documentation:
https://developer.arm.com/documentation/den0028/latest

Change-Id: Id76a2f0c269c8d21f0589415eae1625e713094b6
Signed-off-by: Sona Mathew <SonaRebecca.Mathew@arm.com>