fix(runtime/rmi): update RIPAS for device memory
In update_ripas() function:
- return -EINVAL error for changing RIPAS_DEV to RIPAS_RAM
as not allowed for device mappings
- add support for changing RIPAS_DEV to RIPAS_EMPTY
- remove check for "change_destroyed == CHANGE_DESTROYED"
for changing RIPAS_DESTROYED to RIPAS_EMPTY, as this
flag can be only applied for when changing RIPAS_DESTROYED
to RIPAS_RAM (RMM Specifiaction 1.1-alp12, p. 548).
In smc_dev_mem_map() handler:
- change error condition from !s2tte_is_unassigned() to
!(s2tte_is_unassigned_empty() ||
  s2tte_is_unassigned_destroyed())
to prevent RIPAS_RAM mapping for device memory.

Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>
Change-Id: Iba22c4108fa6b249ab8d6556446b0f44a71ecbfe