refactor(app/common): omit print service using shared
This commit refactors the print el0 app service so that it doesn't use
shared memory.

The print service used the first byte of the shared buffer to pass the
character to print. If the print service was used before the app banked
the first byte, the app might have read the first byte corrupted by the
print service.

Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>
Change-Id: I77189428711de5e1b25b3f6d6ec4acc49f919517