feat(mec): modify XLAT tables to support FEAT_MEC
This patch adds the necessary changes to the XLAT table library
in order to support FEAT_MEC. Specifically, it adds support for
the alternate MECIDs for TTBR1 by setting the proper bits in
register TCR2_EL2. This way, access to RTT, RTT2 and REALM
DATA will be done using MECID_A1_EL2, while access to the
rest of the granules will be done using MECID_P1_EL2.

Change-Id: I6a8fb60dbcb20e8061072c8e015cb26aecc98124
Signed-off-by: Tushar Khandelwal <tushar.khandelwal@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>