feat(mec): enable FEAT_MEC for default MECID
Enable minimum functionality for FEAT_MEC. RMM and Realms MECID is 0
(default MECID). This will be modified in a further so multiple
MECIDs can be supported.

Use MECID 0 as the default because it simplifies RMM setup: we start
with SCRLR2_EL2.EMEC == 0, so the initial MECID is 0, which corresponds
to the MECID used by TF-A for installing RMM code. Then in our own time
we can configure MECID_Px_EL2 and SCTLR2_EL2.EMEC.

Change-Id: I09033a59cffc42762dc8b55d98ac051397adc7e7
Signed-off-by: Tushar Khandelwal <tushar.khandelwal@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>