feat(mec): add RMI calls to handle shared MECID
With this new addtion, the host can set a MECID as shared (provided
it is not being used already) or private.

Else, introduce a new lib/arch component called `mec` and add
necessary helper functions.

In order to preserve default MECID 0 exclusively for RMM code and
data and to prevent Realms from using it, MECIDs assigned to Realms
are incremented by RESERVED_IDS (currently 1). For example, if the
host assigns MECID 0 to a Realm, RMM will program MECID registers
for the realm with value 1; if host assigns MECID 1 to a realm,
Realm will have MECID 2 programmed into its MECID registers, and so
on. Therefore, the range allowed for Realm MECIDs is now between 0
and (2^(MECIDR_EL2.MECIDWidthm1 + 1) - 1 - RESERVED_IDS).

Change-Id: I85d19b2fd92fa6a15e674aa38f859d0d6688e1cd
Signed-off-by: Tushar Khandelwal <tushar.khandelwal@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Soby Mathew <soby.mathew@arm.com>