feat(lib/slot_buf): add ns_buffer_write_unaligned
Add a new function similar to ns_buffer_write that writes to a
non-secure buffer. The new function doesn't mandate 8 byte alignment of
source buffer and length, however ns buf offset alignment is still a
requirement. The new function still uses the optimised aligned memcopy.
If src is not aligned, the unaligned part is copied at an offset, and
the extra bytes are filled with 0. The offset is returned.

Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>
Change-Id: I2a1954ac77696c68ec23ae21f8be19b5625df6a6