feat(mec): make MEC a lib component
It is more useful for MEC component to be able to invoke
rmm-el3-ifc smc calls rather than the user of MEC API.
This cannot be done while MEC is inside the lib/arch folder
as it will introduce cyclic dependancy. This patch moves MEC
out of lib/arch to its own folder within lib.

The patch also modifies mec_reserve() function to make the
call to rmm_el3_ifc_mecid_update(). This means the mec_is_shared()
API is not longer needed and hence it is removed.

Change-Id: Id76bbdab9c61db65be2d10cf2ddd801662a7168a
Signed-off-by: Soby Mathew <soby.mathew@arm.com>