feat(mec): Provide options for granule sanitize
This patch implements alternative granule sanitization methods
as discussed in the FEAT_MEC Design Guide [1]. A new configuration
option, RMM_MEM_SCRUB_METHOD, has been introduced to select between
different sanitize/scrub methods. By default, the granule is zeroed
using the RMM MECID before it is undelegated.

If RMM_MEM_SCRUB_METHOD=1 is selected, a sequentially increasing
counter—starting from a random global seed—is written to the granule
for sanitization.

If RMM_MEM_SCRUB_METHOD=2 is selected, a dedicated MECID is
reserved for scrub purposes. RMM switches to this MECID and zeroes
the granule. Note that this method reserves an additional MECID,
which is therefore unavailable to the NS Host when creating a Realm.
Additionally, if FEAT_MEC is unavailable on the platform, this
method becomes equivalent to the default scrub method.

Additional scrub methods may be added in the future based on
security requirements.

This patch also removes the SLOT_DELEGATED enumeration from
slot_buffer. This change is made because the appropriate
buffer mapping slot—RMM or Realm—depends on the selected
scrub method. The code now consistently uses SLOT_RD for the
RMM slot.

[1] https://github.com/TF-RMM/tf-rmm/wiki/RFC:-FEAT_MEC-Design-in-RMM

Change-Id: Ifebebba6509dffc9c199f015517e34a133087b2d
Signed-off-by: Soby Mathew <soby.mathew@arm.com>