feat(mec): program mec registers in a fine grained manner
The previous patch series introduces the MEC features but programmed
globally on a CPU whenever the Realm Descriptor is mapped to
Stage 1 of RMM. Although functionally this is OK, it is better
to program the mecid registers in a more fine grained manner
as it makes auditing the code base easier. This patch thus makes the
necessary changes to program mecid regs whenever a buffer requiring
Realm MECID is mapped. A new `buffer_granule_mecid_map()` is introduced
for this purpose. The `buffer_unmap()` detects the type of slot used
and reset the mecid appropriately.

Some API names have been refactored to cater for the change in mecid
register programming.

Change-Id: Ic9c8ea9077ae1fe5e079be2f231e72594e17084a
Signed-off-by: Soby Mathew <soby.mathew@arm.com>