feat(mec): fix buffer slot library for fake host
With the introduction of fine-grained MECID programming, the fake
host platform stopped working as the introduced `va_to_slot()` function
would not work due to the way VAs are handled by the fake host.

This patch modifies the slot buffer library to allow the fake host
config to handle VA to slot conversions as well. Suitable host_harness
is introduced to allow host_build, host_test and host_cbmc platforms
to implement suitably. The slot_buffer unit tests are fixed up and
also a test, which tried to unmap a buffer without mapping is
commented out. This test should be coverted to an ASSERT_TEST() as
this is an invalid scenario.

Suitable stubs for CBMC are implemented for `mec` component.
Also the host mock layer now exposes the FEAT_MEC and registers
proper sysreg handlers for the mecid registers.

Change-Id: I5781e7f9196fa3e5f7efcfe672a4550a51c2c0bd
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Soby Mathew <soby.mathew@arm.com>