fix(mec): Introduce new buffer helper for sanitize
When RMM_MEM_SCRUB_METHOD = 2 is configured, using
buffer_map_with_mecid() will not program the RESERVED_SCRUB_MECID
but rather the program a MECID with an offset. This patch introduces
a new helper for granule_sanitize in the slot_buf component and also
some new helpers for programming the RESERVED_SCRUB_MECID which
bypasses the offset calculation.

The new helper, buffer_granule_sanitize(), will perform the
sanitization according to the sanitize policy configured for RMM.

Change-Id: I4712788c03e453949b81216b8bc14aa4d73661df
Signed-off-by: Soby Mathew <soby.mathew@arm.com>