fix(planes): fix some bugs on SIMD trap management for planes
SIMD traps for Plane N are not properly managed. The expected behaviour
should be to propagate CPTR_EL2 register from Plane 0 to Plane N and
viceversa whenever entering or exiting Plane N with TRAP_SIMD flag
disabled so to avoid unnecessary SIMD exceptions when the REC context
has been initialized and Plane N doesn't need to trap the
exception to P0.

In additon to that, another bug which caused that in some instances
the current plane might not be properly updated when exiting plane N,
leaving a stale rec_plane structure that might be then passed to
some APIs has also been fixed.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: Ia04bde406488f85878b26547ec76d5a24bc5b410