fix(mec): Change Realm MEC Policy key value
The patch correct MEC policy value from 44241 to 44243 based on
the ARM Realm Management Monitor Specification v1.1.

The initial value of 44241 was incorrect as per the specification.
This was causing the claim to appear erroneously in the attestation
statement. With this change, the attestation statement includes the
MEC Policy value.

Change-Id: Ia6bb4d2a47290d19b37360de1754eb6a683687a6
Signed-off-by: Andrea Sanchez <andreasanchezp@google.com>