feat(lib/xlat): estimate tables needed for regions
This patch introduces a helper to estimate the number of L2 and L3
tables needed for a set of mmap regions specified by the caller.
There are some assumptions on the mmap regions like they are sorted
in ascending VA and that the regions do not overlap.

This will be used to estimate the number of xlat tables required
at runtime and reserve the same from EL3.

Change-Id: Icdda25b91a43bc16d6f8ddd998c34c77933c95d4
Signed-off-by: Soby Mathew <soby.mathew@arm.com>