feat(lib/xlat): introduce xlat stitch capability
This patch introduces helper to stitch a child xlat table to the
parent xlat table at Level 1. The child xlat table should cover a
region which is a perfect match to block entries in parent table.

This capability will be used to stitch the static and dynamic Low
VA region tables at runtime.

Change-Id: Ie4f1466a03ffdfbc97cc8a262dc7e58aab8d3585
Signed-off-by: Soby Mathew <soby.mathew@arm.com>