feat(lib/xlat): add capability to walk remapped tables
So far, RMM used to have only 1:1 mapping for its low VA space. This
means that the PA of the xlat tables and the VA to access them were
the same. When the xlat tables are mapped at a different VA compared
to the PA, then an appropriate va_to_pa offset needs to be applied
when walking the tables.

If the RMM ensures that the xlat tables for a context are allocated
as a single contigous block of memory, then a constant offset can
be applied for va_to_pa to access the remapped tables when doing a
software walk at runtime.

This patch introduces a new field `tbls_va_to_pa_diff` in the
xlat_ctx_tbls structure and this value will be used during the software
walk.

A new function xlat_ctx_init_remapped_tbls() is introduced to specify
this additional argument for contexts which have tables that are
remapped in the VA space.

Change-Id: Iea5d0984535547d483a162e3e6ae14c103c6b4f0
Signed-off-by: Soby Mathew <soby.mathew@arm.com>