chore(lib/arch): rename hcr_el2 defines
There were two sets of HCR_EL2 values programmed in RMM.
One, called HCR_EL2_INIT, was used during initialization
and also when running EL0 apps. The other, called
HCR_REALM_FLAGS, was used when entering a Realm.

This patch renames HCR_EL2_INIT to HCR_EL2_APP and repurposes
it to be used only when entering EL0 apps. The other flag is
renamed to HCR_EL2_REALM, which is now used during initialization
as well as when entering a Realm.

The comment in the stage2_tlbi_ipa_per_vmids() definition
is corrected to reflect the correct status of {E2H, TGE}
bits in hcr_el2.

Change-Id: I1a47cb0b99326726e1e32e6e024632e6ed7d686e
Signed-off-by: Soby Mathew <soby.mathew@arm.com>