fix(arch): RMM may report wrong PAS size when 52-bit PAS is supported.
RMM considers that FEAT_LPA2 is available whenever 52-bit PAS is
supported on any granularity. This can cause problems if 4KB
granularity doesn't support 52-bit PAS whereas it is supported by
other granularities.

This patch limits the PAS size when 52-bit PAS is not supported for
4KB granularity to a maximum of 48 bits, regardless of whether
FEAT_LPA2 is reported as available or not.

Change-Id: I5382261c747ab7e8f4830ff598422dd8a77eeb01
Signed-off-by: Tuan Chuong Goh <chuong.goh@arm.com>