refactor(plat/host): allow register callback substitution
This patch allows for substitution of the callbacks associated
with reads and writes of a system register. This way, callbacks
can be changed temporarily (or permanently) to observe certain
behaviors.

Previous versions only allowed to set the callbacks once, and the
only way to reset those callbacks was by zeroing all the system
registers and resetting the callback counter, so this addition
improves the flexibility of the host platform.

Change-Id: I1941cb38c59986e3fbb68dbc33bcac92a3d5fe2d
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>