refactor(plat/host): back up installed callbacks index
When emulated system registers are restored, the index pointing at
the location of the array where the next callback should be added is
not restored. Therefore, setting the callbacks for a system register
after a restoration will continue increasing the index until it
falls out of range. This not only is not the intended behavior, but
it prevents from setting a callback once the index is out of range.
This patch solves the problem by backing up the index during
snapshots of the system registers and restoring it during
restoration of the system registers.

Change-Id: I7ab9b017eb56f0b87726f3cb8d0b4bd709b56d49
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>