fix(app): Attestation App Thread Leak
When a REC is created with fake_host, it spawns attestation
app instances, but when it gets destroyed, the REC does not
clean up the apps, leading to a thread leak if multiple
realms/RECs are created.

This patch cleans up the app threads associated with the REC
that is getting destroyed, ignoring the app instances
associated with the main RMM process. Additionally, PDEV
cleans up the associated device assignment apps upon
destruction.

Change-Id: If090a893e91b014b13b4c5df82323d8007eaefbd
Signed-off-by: Tuan Chuong Goh <chuong.goh@arm.com>