feat(firme): use FIRME GPI_SET when available
This patch adds a check at start up where RMM will query the FIRME
interface to see if GPI_SET is supported, and if so it will use
that in place of the delegate and undelegate SMC calls for granule
transitions.

This is based on alp2 version of FIRME specification.

Related TF-A patch:
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/46445

Signed-off-by: John Powell <john.powell@arm.com>
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I41b461246d0a74838c4b1dce648472893fb55493